Breaking Serialization Refactoring - Remove UserObject from the system - Make Serializable objects as light as possible - There is a little performance hit that needs fixing due to the creation of ObjectLayout many times - There is a memory leak in GetIterator, should delete the iterator

+++ ObjectSerializer/Source/TypeResolver.cpp
@@ -105,7 +105,7 @@ void TypeResolver::TypeMemberSubstitution( TypeData& p_tableEntry, TypeTable& p_
     TypeNode*           resultant   = NULL;
     vector<TypeChild>&  oldChildren = typeGraph->Children;
 
-    for(int i = 0; i < oldChildren.size(); ++i)
+    for(size_t i = 0; i < oldChildren.size(); ++i)
     {
         if(!oldChildren[i].IsType)
             continue;
